requirements.txt
# Lichess Friction Analysis - Dependencies
# Core
python-chess>=1.9
pandas>=2.0
numpy>=1.24

# Statistics
scipy>=1.10
statsmodels>=0.14

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Data handling
zstandard>=0.21
pyyaml>=6.0

# Progress
tqdm>=4.65
